refactor: Fix typing error in GameSessionController file name
